refactor: 循環参照を防ぐために実装とタイプでファイルを分割する
