+++ src/ducks/auth.js
@@ -1,7 +1,7 @@
 // @flow
 import firebase from 'firebase';
 
-import type { Dispatch, GetStore } from './';
+import type { Dispatch, GetStore } from './type';
 
 // 最終的な Root Reducere の中で、ここで管理している State が格納される名前
 export const storeName: string = 'auth';
