+++ src/ducks/work.js
@@ -7,7 +7,7 @@ import * as helpers from './helpers';
 import * as auth from './auth';
 import type { Statefull } from './helpers';
 import type { UserType } from './user';
-import type { Dispatch, GetStore } from './';
+import type { Dispatch, GetStore } from './type';
 
 // 最終的な Root Reducere の中で、ここで管理している State が格納される名前
 export const storeName: string = 'work';
